Добавляет разделение на модули
